Work in progress, good start at building broad model
