+++ .gitignore
@@ -165,3 +165,6 @@ cython_debug/
 *.tif
 *.xml
 *.gpkg-*
+*.gpkg
+*.csv
+*.joblib
